backend/trading-service/requirements.txt
# FastAPI和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Redis相关
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 其他
python-dotenv==1.0.0
typing-extensions==4.8.0